Export the key share trail in rageshake (opt-in)
